Changes by Samuel V11

+++ src/main/java/co/edu/cue/proyectoNuclear/mapping/dtos/HistoryStudentDto.java
@@ -3,7 +3,7 @@ package co.edu.cue.proyectoNuclear.mapping.dtos;
 import co.edu.cue.proyectoNuclear.domain.entities.Student;
 import co.edu.cue.proyectoNuclear.domain.entities.Subject;
 
-public record HistoryStudentDto (int id,
+public record HistoryStudentDto (Long id,
                                  Student student,
                                  Subject subject){
 }
